Create users for blog tests also
Blogs are always related to a user
