+++ part4/bloglist/tests/users_api.test.js
@@ -6,20 +6,8 @@ const api = supertest(app);
 const mongo = require("../utils/mongo");
 const config = require("./config");
 const initialUsers = config.mockUsers;
-const { USERS_API_URL, PASSWORD_SALT_ROUNDS } = require("../common/constants");
-
-async function sendUsers(users) {
-    const usersCopy = JSON.parse(JSON.stringify(users)); // Create deep copy
-
-    // Hash the password to passwordHash field
-    for (let i = 0; i < usersCopy.length; i++) {
-        const password = usersCopy[i].password;
-        const passwordHash = await bcrypt.hash(password, PASSWORD_SALT_ROUNDS);
-        delete usersCopy[i].password;
-        usersCopy[i].passwordHash = passwordHash;
-    }
-    await mongo.saveListOfUsers(usersCopy);
-}
+const { sendUsers } = require("./common");
+const { USERS_API_URL } = require("../common/constants");
 
 beforeEach(async () => {
     await mongo.connectDatabase();
